backend/requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
pytesseract
pdf2image
google-generativeai